Refactor ordering method. Move eager loading related methods to EloquentEninge.
